+++ sklearn/datasets/svmlight_format.py
@@ -20,7 +20,7 @@ from ._svmlight_format import _load_svmlight_file
 
 
 def load_svmlight_file(f, n_features=None, dtype=np.float64,
-                       multilabel=False):
+                       multilabel=False, zero_based="auto"):
     """Load datasets in the svmlight / libsvm format into sparse CSR matrix
 
     This format is a text-based format, with one sample per line. It does
@@ -60,6 +60,13 @@ def load_svmlight_file(f, n_features=None, dtype=np.float64,
         Samples may have several labels each (see
         http://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/multilabel.html)
 
+    zero_based: boolean or "auto", optional
+        Whether column indices in f are zero-based (True) or one-based
+        (False). If set to "auto", a heuristic check is applied to determine
+        this from the file contents. Both kinds of files occur "in the wild",
+        but they are unfortunately not self-identifying. Using "auto" or True
+        should always be safe.
+
     Returns
     -------
     (X, y)
@@ -69,13 +76,15 @@ def load_svmlight_file(f, n_features=None, dtype=np.float64,
           a list of tuples of length n_samples.
     """
     if hasattr(f, "read"):
-        return _load_svmlight_file(f, n_features, dtype, multilabel)
+        return _load_svmlight_file(f, n_features, dtype, multilabel,
+                                   zero_based)
     with open(f, 'rb') as f:
-        return _load_svmlight_file(f, n_features, dtype, multilabel)
+        return _load_svmlight_file(f, n_features, dtype, multilabel,
+                                   zero_based)
 
 
 def load_svmlight_files(files, n_features=None, dtype=np.float64,
-                        multilabel=False):
+                        multilabel=False, zero_based=True):
     """Load dataset from multiple files in SVMlight format
 
     This function is equivalent to mapping load_svmlight_file over a list of
@@ -99,6 +108,10 @@ def load_svmlight_files(files, n_features=None, dtype=np.float64,
         Samples may have several labels each (see
         http://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/multilabel.html)
 
+    zero_based: boolean, optional
+        Whether column indices in files are zero-based (True) or one-based
+        (False).
+
     Returns
     -------
     [X1, y1, ..., Xn, yn]
@@ -116,30 +129,36 @@ def load_svmlight_files(files, n_features=None, dtype=np.float64,
     --------
     load_svmlight_file
     """
+    if zero_based == "auto":
+        raise TypeError("zero_based must be boolean; the heuristic"
+                        " implemented by load_svmlight_file is not supported")
     files = iter(files)
-    result = list(load_svmlight_file(files.next(), n_features, dtype))
+    result = list(load_svmlight_file(files.next(), n_features, dtype,
+                                     multilabel, zero_based))
     n_features = result[0].shape[1]
 
     for f in files:
-        result += load_svmlight_file(f, n_features, dtype, multilabel)
+        result += load_svmlight_file(f, n_features, dtype, multilabel,
+                                     zero_based)
 
     return result
 
 
-def _dump_svmlight(X, y, f):
+def _dump_svmlight(X, y, f, zero_based):
     if X.shape[0] != y.shape[0]:
         raise ValueError("X.shape[0] and y.shape[0] should be the same, "
                          "got: %r and %r instead." % (X.shape[0], y.shape[0]))
 
     is_sp = int(hasattr(X, "tocsr"))
 
+    one_based = not zero_based
     for i in xrange(X.shape[0]):
-        s = u" ".join([u"%d:%f" % (j + 1, X[i, j])
+        s = u" ".join([u"%d:%f" % (j + one_based, X[i, j])
                        for j in X[i].nonzero()[is_sp]])
         f.write((u"%f %s\n" % (y[i], s)).encode('ascii'))
 
 
-def dump_svmlight_file(X, y, f):
+def dump_svmlight_file(X, y, f, zero_based=True):
     """Dump the dataset in svmlight / libsvm file format.
 
     This format is a text-based format, with one sample per line. It does
@@ -160,9 +179,13 @@ def dump_svmlight_file(X, y, f):
     f : str or file-like in binary mode
         If string it specifies the path that will contain the data.
         If f is a file-like then data will be written to f.
+
+    zero_based : boolean, optional
+        Whether column indices should be written zero-based (True) or one-based
+        (False).
     """
     if hasattr(f, "write"):
-        _dump_svmlight(X, y, f)
+        _dump_svmlight(X, y, f, zero_based)
     else:
         with open(f, "wb") as f:
-            _dump_svmlight(X, y, f)
+            _dump_svmlight(X, y, f, zero_based)